Remove Winsock interop files from the *nix builds.
