+++ src/System.Net.Sockets/src/System/Net/Sockets/SendPacketsElementFlags.Windows.cs
@@ -1,0 +1,12 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace System.Net.Sockets
+{
+    internal enum SendPacketsElementFlags
+    {
+        File = Interop.Winsock.TransmitPacketsElementFlags.File,
+        Memory = Interop.Winsock.TransmitPacketsElementFlags.Memory,
+        EndOfPacket = Interop.Winsock.TransmitPacketsElementFlags.EndOfPacket,
+    }
+}
