+++ src/System.Net.Sockets/src/System/Net/Sockets/Socket.Windows.cs
@@ -17,6 +17,114 @@ namespace System.Net.Sockets
 {
     public partial class Socket
     {
+        private DynamicWinsockMethods _dynamicWinsockMethods;
+
+        private void EnsureDynamicWinsockMethods()
+        {
+            if (_dynamicWinsockMethods == null)
+            {
+                _dynamicWinsockMethods = DynamicWinsockMethods.GetMethods(_addressFamily, _socketType, _protocolType);
+            }
+        }
+
+        internal bool AcceptEx(SafeCloseSocket listenSocketHandle,
+                              SafeCloseSocket acceptSocketHandle,
+                              IntPtr buffer,
+                              int len,
+                              int localAddressLength,
+                              int remoteAddressLength,
+                              out int bytesReceived,
+                              SafeHandle overlapped)
+        {
+            EnsureDynamicWinsockMethods();
+            AcceptExDelegate acceptEx = _dynamicWinsockMethods.GetDelegate<AcceptExDelegate>(listenSocketHandle);
+
+            return acceptEx(listenSocketHandle,
+                            acceptSocketHandle,
+                            buffer,
+                            len,
+                            localAddressLength,
+                            remoteAddressLength,
+                            out bytesReceived,
+                            overlapped);
+        }
+
+        internal void GetAcceptExSockaddrs(IntPtr buffer,
+                                           int receiveDataLength,
+                                           int localAddressLength,
+                                           int remoteAddressLength,
+                                           out IntPtr localSocketAddress,
+                                           out int localSocketAddressLength,
+                                           out IntPtr remoteSocketAddress,
+                                           out int remoteSocketAddressLength)
+        {
+            EnsureDynamicWinsockMethods();
+            GetAcceptExSockaddrsDelegate getAcceptExSockaddrs = _dynamicWinsockMethods.GetDelegate<GetAcceptExSockaddrsDelegate>(_handle);
+
+            getAcceptExSockaddrs(buffer,
+                                 receiveDataLength,
+                                 localAddressLength,
+                                 remoteAddressLength,
+                                 out localSocketAddress,
+                                 out localSocketAddressLength,
+                                 out remoteSocketAddress,
+                                 out remoteSocketAddressLength);
+        }
+
+        internal bool DisconnectEx(SafeCloseSocket socketHandle, SafeHandle overlapped, int flags, int reserved)
+        {
+            EnsureDynamicWinsockMethods();
+            DisconnectExDelegate disconnectEx = _dynamicWinsockMethods.GetDelegate<DisconnectExDelegate>(socketHandle);
+
+            return disconnectEx(socketHandle, overlapped, flags, reserved);
+        }
+
+        private bool DisconnectEx_Blocking(IntPtr socketHandle, IntPtr overlapped, int flags, int reserved)
+        {
+            EnsureDynamicWinsockMethods();
+            DisconnectExDelegate_Blocking disconnectEx_Blocking = _dynamicWinsockMethods.GetDelegate<DisconnectExDelegate_Blocking>(_handle);
+
+            return disconnectEx_Blocking(socketHandle, overlapped, flags, reserved);
+        }
+
+        internal bool ConnectEx(SafeCloseSocket socketHandle,
+                               IntPtr socketAddress,
+                               int socketAddressSize,
+                               IntPtr buffer,
+                               int dataLength,
+                               out int bytesSent,
+                               SafeHandle overlapped)
+        {
+            EnsureDynamicWinsockMethods();
+            ConnectExDelegate connectEx = _dynamicWinsockMethods.GetDelegate<ConnectExDelegate>(socketHandle);
+
+            return connectEx(socketHandle, socketAddress, socketAddressSize, buffer, dataLength, out bytesSent, overlapped);
+        }
+
+        internal SocketError WSARecvMsg(SafeCloseSocket socketHandle, IntPtr msg, out int bytesTransferred, SafeHandle overlapped, IntPtr completionRoutine)
+        {
+            EnsureDynamicWinsockMethods();
+            WSARecvMsgDelegate recvMsg = _dynamicWinsockMethods.GetDelegate<WSARecvMsgDelegate>(socketHandle);
+
+            return recvMsg(socketHandle, msg, out bytesTransferred, overlapped, completionRoutine);
+        }
+
+        internal SocketError WSARecvMsg_Blocking(IntPtr socketHandle, IntPtr msg, out int bytesTransferred, IntPtr overlapped, IntPtr completionRoutine)
+        {
+            EnsureDynamicWinsockMethods();
+            WSARecvMsgDelegate_Blocking recvMsg_Blocking = _dynamicWinsockMethods.GetDelegate<WSARecvMsgDelegate_Blocking>(_handle);
+
+            return recvMsg_Blocking(socketHandle, msg, out bytesTransferred, overlapped, completionRoutine);
+        }
+
+        internal bool TransmitPackets(SafeCloseSocket socketHandle, IntPtr packetArray, int elementCount, int sendSize, SafeNativeOverlapped overlapped, TransmitFileOptions flags)
+        {
+            EnsureDynamicWinsockMethods();
+            TransmitPacketsDelegate transmitPackets = _dynamicWinsockMethods.GetDelegate<TransmitPacketsDelegate>(socketHandle);
+
+            return transmitPackets(socketHandle, packetArray, elementCount, sendSize, overlapped, flags);
+        }
+
         internal static IntPtr[] SocketListToFileDescriptorSet(IList socketList)
         {
             if (socketList == null || socketList.Count == 0)
